test: move unified MRP coverage into integration suite
