+++ .github/workflows/ci.yml
@@ -104,4 +104,4 @@ jobs:
           pip install -e ".[dev]"
 
       - name: Run integration tests (MRP endpoints)
-        run: python -m pytest tests/integration/test_mrp_api.py -v --tb=short
+        run: python -m pytest tests/integration/test_mrp_api.py tests/integration/test_mrp_unified.py -v --tb=short
